Add dummy quote implementation for asset messages
